requirements.txt
streamlit==1.37.0
openpyxl==3.1.5
boto3==1.34.149
python_dotenv==1.0.1
snowflake-connector-python